re: 내정보 수정하기

+++ src/Repositories/userRepository.js
@@ -34,7 +34,7 @@ export const findUserByEmail = async (email) => {
 };
 
 // 사용자 정보 수정기능
-export const userModify = async (data)=>{
+export const userPositionModify = async (data)=>{
     const connection = await pool.getConnection();
     try{
         // 사용자 역할 수정
@@ -42,32 +42,37 @@ export const userModify = async (data)=>{
             `update USERS set position = "${data.position}" where id = ${data.userId};`,
         )
 
-        const skills = data.skill.map(skill => `SELECT '${skill}'`).join(' union all ');
-        const fromSQL = `(${skills}) AS new_techs(tech_name)`;
-
-
-        //기술스택 추가하기
-        const [result2] = await connection.query(
-            `insert into USER_TECH_STACKS (user_id, tech_name)
-            select ${data.userId}, tech_name
-            from ${fromSQL}
-            where tech_name NOT IN (
-                select tech_name
-                from USER_TECH_STACKS
-                where user_id = ${data.userId}
-            );`
-        )
-
-        const mySkill = data.skill.map(skill => `'${skill}'`).join(', ');
-
-        // 필요없는 기술 스택 삭제 (deleteStack에 등록된 기술스택 제외하고 다 삭제)
-        const [deleteStackResults] = await connection.query(
-            `delete from USER_TECH_STACKS 
-            where user_id = ${data.userId} and tech_name not in (${mySkill});`
-        )
         return data.userId;
     }catch(err){
         throw new Error(`오류 발생 파라미터 확인바람 (${err})`)
+    }finally {
+        connection.release();
+    }
+}
+
+// 기술스택 수정
+export const userSkillModify = async (userId, skill)=>{
+    const connection = await pool.getConnection();
+    try{
+        await connection.query(
+            `insert into USER_TECH_STACKS (user_id, tech_name) values(${userId}, '${skill}');`
+        );
+    }catch(err){
+        throw new Error(`오류 발생 파라미터 확인바람 (${err})`)
+    }finally {
+        connection.release();
+    }
+}
+
+// 해당 유저의 기술스택 전부 삭제
+export const userSkillDel = async (userId)=>{
+    const connection = await pool.getConnection();
+    try{
+        await connection.query(
+            `delete from USER_TECH_STACKS where user_id = ${userId};`
+        )
+    }catch(err){
+        throw new Error(`오류 발생 파라미터 확인 (${err})`)
     }finally {
         connection.release();
     }
